% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getResults}
\alias{getResults}
\title{Internal function to ggregate results into the final matrix.}
\usage{
getResults(
  printTwoSided = 1L,
  usePrintAlt = 0L,
  usePrintCnew = 0L,
  usePrintID = 0L,
  usePrintAn = 1L
)
}
\arguments{
\item{printTwoSided}{print two sided p-values}

\item{usePrintAlt}{print also alternative side, i.e. depletion}

\item{usePrintCnew}{print old community ids}

\item{usePrintID}{print annotation ID if \code{TRUE},
else annotation description}

\item{usePrintAn}{print annotation type size}
}
\description{
!Note! This internal function is not for regular users.
Use \code{\link{run_enrichment}} for analysis.
}
