% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load}
\alias{load}
\title{Internal function to load data for calculation.}
\usage{
load(x, anno)
}
\arguments{
\item{x}{group/cluster membership data with element name in the
first column and group id in the second}

\item{anno}{element annotation data with term id, term name and
the element name in columns one, two and three respectively.}
}
\description{
This function reads group/cluster membership and element
annotation data for analysis. It does not return anything,
it just create required structures in memory.
}
\details{
!Note! This internal function is not for regular users.
Use \code{\link{run_enrichment}} for analysis.
}
