% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEnrich.R
\name{makeDF}
\alias{makeDF}
\title{Convert result matrix to data.frame.}
\usage{
makeDF(tt, terms, gname, N, usePrintAlt)
}
\arguments{
\item{tt}{result matrix from \code{\link{getResults}}}

\item{terms}{list of unique annotation terms}

\item{gname}{name of grouping}

\item{N}{size of the universe}

\item{usePrintAlt}{}
}
\value{
data.frame with the following columns:
"Fn","C","Cn","N","Mu","OR","CIl","CIu","Pv","Ap","PvALT","ApALT",
"Fe","E(Mu)","F","Alg","Fc"
}
\description{
Creates data.frame with the following columns:
\itemize{
\item Fn annotation term ID
\item C cluster/group ID
\item Cn size of the cluster
\item N size of the universe
\item Mu actual number of term found in cluster
\item OR
\item CIl lower CI boundary
\item CIu upper CI boundary
\item Pv p.value
\item Ap adjusted p.value
\item PvALT p-value for the alternative side, i.e. depletion
if \code{\link{usePrintAlt}} is set to \code{TRUE}.
\item ApALT adjusted p.value for the alternative side, i.e. depletion
if \code{\link{usePrintAlt}} is set to \code{TRUE}.
\item Fe
\item E(Mu) expected number of term found in cluster/group by chance
\item F total number of term instances in the universe
\item Alg name of grouping algorithm/approach
\item Fc
}
}
