% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run}
\alias{run}
\title{Internal function to run enrichment analysis.}
\usage{
run(
  useChi2 = 0L,
  useOneSided = 0L,
  useTwoSided = 1L,
  useMaxSS = 0L,
  runPerm = 0L,
  singlePerm = 0L,
  useSeed = 0L,
  setNOP = 0L,
  pesudoCount = -1,
  FDRmeth = "BY"
)
}
\arguments{
\item{useChi2}{=0,}

\item{useOneSided}{=0,}

\item{useTwoSided}{=1,}

\item{useMaxSS}{=0,}

\item{runPerm}{=0,}

\item{singlePerm}{=0,}

\item{useSeed}{=0,}

\item{setNOP}{=0,}

\item{pesudoCount}{=-1.0,}

\item{FDRmeth}{="BY"}
}
\description{
This function perform actual calculations.
}
\details{
!Note! This internal function is not for regular users.
Use \code{\link{run_enrichment}} for analysis.
}
