% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEnrich.R
\name{run_enrichment}
\alias{run_enrichment}
\title{Run enrichment analysis}
\usage{
run_enrichment(
  membership,
  annotation,
  gname,
  printTwoSided = TRUE,
  usePrintAlt = TRUE,
  usePrintID = TRUE
)
}
\arguments{
\item{membership}{membership data.frame with columns \code{names} and
\code{membership}}

\item{annotation}{annotation data.frame with columns \code{names},
\code{termID} and \code{termName} for each object-term
pair.}

\item{gname}{name of the grouping to distinguish from other membership data.
For example if membership generated by graph clustering algorithm,
name of this algorithm should be used to compare enrichment results
between algorithms.}

\item{printTwoSided}{print two-sided p-value}

\item{usePrintAlt}{print also alternative side, i.e. depletion}

\item{usePrintID}{print annotation ID, else annotation description}
}
\value{
return \code{data.frame} with row for earch group-term pair
}
\description{
This function calculate results of enrichment and return \code{data.frame}
for earch group-term pair with the following columns:
\itemize{
\item Fn annotation term ID
\item C cluster/group ID
\item Cn size of the cluster
\item N size of the universe
\item Mu actual number of term found in cluster
\item OR
\item CIl lower CI boundary
\item CIu upper CI boundary
\item Pv p.value
\item Ap adjusted p.value
\item PvALT p-value for the alternative side, i.e. depletion
if \code{usePrintAlt} is set to \code{TRUE}.
\item ApALT adjusted p.value for the alternative side, i.e. depletion
if \code{usePrintAlt} is set to \code{TRUE}.
\item Fe
\item E(Mu) expected number of term found in cluster/group by chance
\item F total number of term instances in the universe
\item Alg name of grouping algorithm/approach
\item Fc
}
}
\examples{
library(igraph)
## load graph
gfile <- system.file("extdata", "PPI_Presynaptic.gml", package = "rEnrich")
gg = read.graph(gfile,format="gml")
## build cluster membership data.frame
ids  = igraph::get.vertex.attribute(gg,"name",V(gg))
coms = igraph::get.vertex.attribute(gg,"louvain",V(gg))
membership = as.data.frame(cbind(ids,coms))
names(membership)<-c('names','membership')
## load annotation flat file
afile <- system.file("extdata", "flatfile_human_gene2HDO.csv", package = "rEnrich")
anno = read.delim(afile,skip=1,sep="\t",header=F)
anno = as.data.frame(anno)
names(anno)<-c('termID','termName','names')
df <- run_enrichment(membership,anno,'louvain')
}
